Moved legacy WebForms files into new project
Updated paths, namespaces. New project depends on Ignia.Topics, but
Ignia.Topics is unaware of WebForms. Removed references to
System.Configuration and System.Web from Ignia.Topics to be certain.
