+++ Ignia.Topics.Web/Configuration/SourceElementCollection.cs
@@ -7,33 +7,34 @@ using System;
 using System.Configuration;
 using System.Diagnostics.Contracts;
 
-namespace Ignia.Topics.Configuration {
+namespace Ignia.Topics.Web.Configuration {
 
   /*============================================================================================================================
-  | CLASS: PAGE TYPE ELEMENT COLLECTION
+  | CLASS: SOURCE ELEMENT COLLECTION
   \---------------------------------------------------------------------------------------------------------------------------*/
   /// <summary>
   ///   Provides a custom implementation of a <see cref="ConfigurationElementCollection"/> responsible for encapsulating a set
-  ///   of <see cref="PageTypeElement"/> elements.
+  ///   of <see cref="SourceElement"/> elements.
   /// </summary>
   /// <remarks>
-  ///   Includes the Default property that exposes the default page type (<see cref="Ignia.Topics.Web.TopicPage"/>).
+  ///   Adapted DIRECTLY from the Ignia Localization library; in the future, these libraries may (and should) share custom
+  ///   configuration classes.
   /// </remarks>
-  public class PageTypeElementCollection : ConfigurationElementCollection {
+  public class SourceElementCollection : ConfigurationElementCollection {
 
     /*==========================================================================================================================
     | INDEXER
     \-------------------------------------------------------------------------------------------------------------------------*/
     /// <summary>
-    ///   Gets or sets the <see cref="PageTypeElement"/> at the specified index.
+    ///   Gets or sets the <see cref="SourceElement"/> at the specified index.
     /// </summary>
-    /// <param name="index">The integer index for the <see cref="PageTypeElement"/> item in the collection.</param>
+    /// <param name="index">The integer index for the <see cref="SourceElement"/> item in the collection.</param>
     /// <requires description="The value from the getter must not be null." exception="T:System.ArgumentNullException">
     ///   value != null
     /// </requires>
-    public PageTypeElement this[int index] {
+    public SourceElement this[int index] {
       get {
-        return base.BaseGet(index) as PageTypeElement;
+        return base.BaseGet(index) as SourceElement;
       }
       set {
         Contract.Requires<ArgumentNullException>(value != null, "The value from the getter must not be null.");
@@ -45,27 +46,14 @@ namespace Ignia.Topics.Configuration {
     }
 
     /*==========================================================================================================================
-    | ATTRIBUTE: DEFAULT
-    \-------------------------------------------------------------------------------------------------------------------------*/
-    /// <summary>
-    ///   Gets the default page type (e.g., <see cref="Ignia.Topics.Web.TopicPage"/>).
-    /// </summary>
-    [ConfigurationProperty("default", DefaultValue="TopicPage", IsRequired = false)]
-    public string Default {
-      get {
-        return (string)base["default"];
-      }
-    }
-
-    /*==========================================================================================================================
     | METHOD: CREATE NEW ELEMENT
     \-------------------------------------------------------------------------------------------------------------------------*/
     /// <summary>
-    ///   Creates a new <see cref="PageTypeElement"/>.
+    ///   Creates a new <see cref="ConfigurationElement"/>.
     /// </summary>
-    /// <returns>A new instance of a <see cref="PageTypeElement"/>.</returns>
+    /// <returns>A new instance of a <see cref="ConfigurationElement"/>.</returns>
     protected override ConfigurationElement CreateNewElement() {
-      return new PageTypeElement();
+      return new SourceElement();
     }
 
     /*==========================================================================================================================
@@ -75,13 +63,13 @@ namespace Ignia.Topics.Configuration {
     ///   Gets the key for the <see cref="ConfigurationElement"/> item in the collection.
     /// </summary>
     /// <param name="element">The <see cref="ConfigurationElement"/> element object from which to extract the key.</param>
-    /// <returns>The Name string value for the <see cref="PageTypeElement"/> as the element's key.</returns>
+    /// <returns>The Source string value for the <see cref="SourceElement"/> as the element's key.</returns>
     protected override object GetElementKey(ConfigurationElement element) {
-      Contract.Assert(
-        ((PageTypeElement)element) != null,
-        "Confirms the element's Name value is available when deriving its key."
+      Contract.Assume(
+        ((SourceElement)element).Source != null,
+        "Confirm the element's Source value is available when deriving its key."
       );
-      return ((PageTypeElement)element);
+      return ((SourceElement)element).Source;
     }
 
   } // Class
