+++ Ignia.Topics.Web/Configuration/VersioningElement.cs
@@ -5,7 +5,7 @@
 \=============================================================================================================================*/
 using System.Configuration;
 
-namespace Ignia.Topics.Configuration {
+namespace Ignia.Topics.Web.Configuration {
 
   /*============================================================================================================================
   | CLASS: VERSIONING ELEMENT
