+++ Ignia.Topics.Web/Editor/AttributeTypeControl.cs
@@ -5,7 +5,7 @@
 \=============================================================================================================================*/
 using System.Web.UI;
 
-namespace Ignia.Topics.Editor {
+namespace Ignia.Topics.Web.Editor {
 
   /*============================================================================================================================
   | CLASS: ATTRIBUTE TYPE CONTROL
