+++ Ignia.Topics.Web/Editor/IEditControl.cs
@@ -4,7 +4,7 @@
 | Project       Topics Library
 \=============================================================================================================================*/
 
-namespace Ignia.Topics.Editor {
+namespace Ignia.Topics.Web.Editor {
 
   /*============================================================================================================================
   | INTERFACE
