+++ Ignia.Topics.sln
@@ -14,6 +14,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ignia.Topics.Data.Sql", "Ig
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ignia.Topics.Data.Caching", "Ignia.Topics.Data.Caching\Ignia.Topics.Data.Caching.csproj", "{206B7F91-CA25-4E9D-9576-60D2E54A2C0A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ignia.Topics.Web", "Ignia.Topics.Web\Ignia.Topics.Web.csproj", "{C98F7B48-A085-4394-B820-C244F23868CE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug CC|Any CPU = Debug CC|Any CPU
@@ -57,6 +59,12 @@ Global
 		{206B7F91-CA25-4E9D-9576-60D2E54A2C0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{206B7F91-CA25-4E9D-9576-60D2E54A2C0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{206B7F91-CA25-4E9D-9576-60D2E54A2C0A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C98F7B48-A085-4394-B820-C244F23868CE}.Debug CC|Any CPU.ActiveCfg = Debug|Any CPU
+		{C98F7B48-A085-4394-B820-C244F23868CE}.Debug CC|Any CPU.Build.0 = Debug|Any CPU
+		{C98F7B48-A085-4394-B820-C244F23868CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C98F7B48-A085-4394-B820-C244F23868CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C98F7B48-A085-4394-B820-C244F23868CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C98F7B48-A085-4394-B820-C244F23868CE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
